projects/2026-02-08/ClawTrade/Crypto_Agent/crypto_quant_agent/requirements.txt
ccxt==4.0.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
mplfinance==0.12.10b0
yfinance==0.2.28
requests==2.31.0
scipy==1.11.1
google-generativeai==0.3.2
flask==3.0.0
flask-cors==4.0.0
